animation: split into components
